Format the covariate effect simulation results for printing

Format the covariate effect simulation results for printing

Usage

print_coveff(
  coveffsim,
  n_sigfig = 3,
  use_seps = TRUE,
  drop_trailing_dec_mark = TRUE
)

Arguments

coveffsim

an object of class coveffsim

n_sigfig

Number of significant figures to form value_label of continuous variables. See gt::vec_fmt_number() for details.

use_seps

Whether to use separators for thousands in printing numbers. See gt::vec_fmt_number() for details.

drop_trailing_dec_mark

Whether to drop the trailing decimal mark (".") in value_label of continuous variables. See gt::vec_fmt_number() for details.

Value

A data frame with the formatted covariate effect simulation results with the following columns:

var_label: the label of the covariate

value_label: the label of the covariate value

value_annot: the annotation of the covariate value

Odds ratio: the odds ratio of the covariate effect

95% CI: the 95% credible interval of the covariate effect

Details

Note that n_sigfig, use_seps, and drop_trailing_dec_mark are only applied to the odds ratio and 95% CI columns; value_label column was already generated in an earlier step in build_spec_coveff() or sim_coveff().

Examples

# \donttest{
data(d_sim_binom_cov_hgly2)

ermod_bin <- dev_ermod_bin(
  data = d_sim_binom_cov_hgly2,
  var_resp = "AEFLAG",
  var_exposure = "AUCss_1000",
  var_cov = "BHBA1C_5",
)

print_coveff(sim_coveff(ermod_bin))
#> # A tibble: 6 × 5
#>   var_label  value_label value_annot `Odds ratio` `95% CI`        
#>   <chr>      <chr>       <chr>       <chr>        <chr>           
#> 1 AUCss_1000 0.868       5th         0.548        "[0.452, 0.652]"
#> 2 AUCss_1000 2.21        median      1            " "             
#> 3 AUCss_1000 5.30        95th        3.98         "[2.67, 6.18]"  
#> 4 BHBA1C_5   5.75        5th         0.276        "[0.200, 0.372]"
#> 5 BHBA1C_5   7.97        median      1            " "             
#> 6 BHBA1C_5   10.4        95th        4.20         "[3.02, 6.00]"  
# }